Created 3 trajectories and ramsete commands
