--- RobotCode2020New/src/main/java/frc/robot/autocommand/LoadedSkilletCommand.java
@@ -1,82 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-package frc.robot.autocommand;
-
-
-import edu.wpi.first.wpilibj.controller.RamseteController;
-
-import edu.wpi.first.wpilibj.trajectory.TrajectoryGenerator;
-import edu.wpi.first.wpilibj.trajectory.Trajectory;
-import edu.wpi.first.wpilibj.trajectory.constraint.DifferentialDriveVoltageConstraint;
-import edu.wpi.first.wpilibj.trajectory.TrajectoryConfig;
-
-import edu.wpi.first.wpilibj.controller.SimpleMotorFeedforward;
-
-import edu.wpi.first.wpilibj.geometry.Rotation2d;
-import edu.wpi.first.wpilibj.geometry.Pose2d;
-import edu.wpi.first.wpilibj.geometry.Translation2d;
-import edu.wpi.first.wpilibj.controller.PIDController;
-
-import edu.wpi.first.wpilibj2.command.*;
-
-import frc.robot.subsystems.DriveSubsystem;
-import frc.robot.Constants;
-
-import java.util.List;
-public class LoadedSkilletCommand extends SequentialCommandGroup {
-  /**
-   * Creates a new Autonomous.
-   */
-  DriveSubsystem m_drive;
-  public LoadedSkilletCommand(DriveSubsystem subsystem) {
-    m_drive = subsystem;
-    var autoVoltageConstraint =
-        new DifferentialDriveVoltageConstraint(
-            new SimpleMotorFeedforward(Constants.KSVOLTS,
-            Constants.KVVOLT,
-            Constants.KAVOLT),
-            Constants.KDRIVEKINEMATICS,10);
-    
-    // Configurate the values of all trajectories for max velocity and acceleration
-    TrajectoryConfig config =
-      new TrajectoryConfig(Constants.KMAXSPEED,
-      Constants.KMAXACCELERATION)
-      // Add kinematics to ensure max speed is actually obeyed
-      .setKinematics(Constants.KDRIVEKINEMATICS)
-      // Apply the voltage constraint
-      .addConstraint(autoVoltageConstraint);
-
-    // Generates a trajectory for a path in autonomous
-    Trajectory trajectory1 = TrajectoryGenerator.generateTrajectory(
-        // Start at the origin facing the +X direction
-        new Pose2d(0, 0, new Rotation2d(0)),
-        List.of(),
-        // End 1 meters straight ahead of where we started, facing forward
-        new Pose2d(1, -1, new Rotation2d(270)),
-        // Pass config
-        config);
-
-    RamseteCommand ramseteCommand1 = new RamseteCommand(
-        trajectory1,
-        m_drive::getPose,
-        new RamseteController(Constants.KRAMSETEB, Constants.KRAMSETEZETA),
-        new SimpleMotorFeedforward(Constants.KSVOLTS,
-                                   Constants.KVVOLT,
-                                   Constants.KAVOLT),
-        Constants.KDRIVEKINEMATICS,
-        m_drive::getWheelSpeeds,
-        new PIDController(Constants.KPDRIVEVEL, 0, 0),
-        new PIDController(Constants.KPDRIVEVEL, 0, 0),
-        // RamseteCommand passes volts to the callback
-        m_drive::tankDriveVolts,
-        m_drive
-    );
-
-    addCommands(ramseteCommand1,new WaitCommand(5));
-  }
-}
