+++ RobotCode2020New/src/main/java/frc/robot/commands/autocommands/CheesyDenverSkilletCommand.java
@@ -1,2 +1,33 @@
+/*----------------------------------------------------------------------------*/
+/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
+/* Open Source Software - may be modified and shared by FRC teams. The code   */
+/* must be accompanied by the FIRST BSD license file in the root directory of */
+/* the project.                                                               */
+/*----------------------------------------------------------------------------*/
 
 package frc.robot.commands.autocommands;
+
+import edu.wpi.first.wpilibj.controller.RamseteController;
+
+import edu.wpi.first.wpilibj.trajectory.TrajectoryGenerator;
+import edu.wpi.first.wpilibj.trajectory.Trajectory;
+import edu.wpi.first.wpilibj.trajectory.constraint.DifferentialDriveVoltageConstraint;
+import edu.wpi.first.wpilibj.trajectory.TrajectoryConfig;
+
+import edu.wpi.first.wpilibj.controller.SimpleMotorFeedforward;
+
+import edu.wpi.first.wpilibj.geometry.Rotation2d;
+import edu.wpi.first.wpilibj.geometry.Pose2d;
+import edu.wpi.first.wpilibj.geometry.Translation2d;
+import edu.wpi.first.wpilibj.controller.PIDController;
+
+import edu.wpi.first.wpilibj2.command.*;
+
+import frc.robot.subsystems.DriveSubsystem;
+import frc.robot.Constants;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class CheesyDenverSkilletCommand extends SequentialCommandGroup {
+}
